mgr/rook: fix osd creation using StorageClassDeviceSets
This commit fixes osd creation by implementing the apply_drivegroups method
for the RookOrchestrator class and the add_osds method for the RookCluster
class.

The apply_drivegroups method, filters matching hosts according to the drive
group spec, and calls the add_osds method.

The add_osds method makes a call to patch the Ceph Cluster CR with StorageClassDeviceSets
to create osds consuming the storage class specified in the ceph config at mgr/rook/storage_class
on the list of matching hosts passed.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>